refactor: update naming and layout
