+++ examples/e2e/app-router/open-next.config.ts
@@ -1,13 +1,13 @@
 import { defineCloudflareConfig } from "@opennextjs/cloudflare";
-import kvIncrementalCache from "@opennextjs/cloudflare/kv-cache";
-import shardedTagCache from "@opennextjs/cloudflare/do-sharded-tag-cache";
-import doQueue from "@opennextjs/cloudflare/durable-queue";
+import kvIncrementalCache from "@opennextjs/cloudflare/overrides/incremental-cache/kv-cache";
+import shardedTagCache from "@opennextjs/cloudflare/overrides/tag-cache/do-sharded-tag-cache";
+import doQueue from "@opennextjs/cloudflare/overrides/queue/do-queue";
 
 export default defineCloudflareConfig({
   incrementalCache: kvIncrementalCache,
-  // With such a configuration, we could have up to 12 * (8 + 2) = 120 Durable Objects instances
+  // With such a configuration, we could have up to 2 * 6 * (8 + 2) = 120 Durable Objects instances
   tagCache: shardedTagCache({
-    numberOfShards: 12,
+    baseShardSize: 6,
     enableShardReplication: true,
     shardReplicationOptions: {
       numberOfSoftReplicas: 8,
