+++ examples/overrides/memory-queue/wrangler.jsonc
@@ -10,7 +10,7 @@
   },
   "kv_namespaces": [
     {
-      "binding": "NEXT_CACHE_WORKERS_KV",
+      "binding": "NEXT_CACHE_KV",
       "id": "<BINDING_ID>"
     }
   ],
