+++ examples/overrides/r2-incremental-cache/open-next.config.ts
@@ -1,8 +1,8 @@
 import { defineCloudflareConfig } from "@opennextjs/cloudflare";
-import d1TagCache from "@opennextjs/cloudflare/d1-tag-cache";
-import memoryQueue from "@opennextjs/cloudflare/memory-queue";
-import r2IncrementalCache from "@opennextjs/cloudflare/r2-incremental-cache";
-import { withRegionalCache } from "@opennextjs/cloudflare/regional-cache";
+import d1TagCache from "@opennextjs/cloudflare/overrides/tag-cache/d1-tag-cache";
+import memoryQueue from "@opennextjs/cloudflare/overrides/queue/memory-queue";
+import r2IncrementalCache from "@opennextjs/cloudflare/overrides/incremental-cache/r2-incremental-cache";
+import { withRegionalCache } from "@opennextjs/cloudflare/overrides/incremental-cache/regional-cache";
 
 export default defineCloudflareConfig({
   incrementalCache: withRegionalCache(r2IncrementalCache, {
