+++ examples/ssg-app/open-next.config.ts
@@ -1,5 +1,5 @@
 import { defineCloudflareConfig } from "@opennextjs/cloudflare";
-import kvIncrementalCache from "@opennextjs/cloudflare/kv-cache";
+import kvIncrementalCache from "@opennextjs/cloudflare/overrides/incremental-cache/kv-cache";
 
 export default defineCloudflareConfig({
   incrementalCache: kvIncrementalCache,
