+++ packages/cloudflare/CHANGELOG.md
@@ -425,7 +425,7 @@
 
 - 87f4fb5: feat: configure kv binding name with env var
 
-  The Workers KV binding used in the Next.js cache handler can be given a custom name with the `__OPENNEXT_KV_BINDING_NAME` environment variable at build-time, instead of defaulting to `NEXT_CACHE_WORKERS_KV`.
+  The Workers KV binding used in the Next.js cache handler can be given a custom name with the `__OPENNEXT_KV_BINDING_NAME` environment variable at build-time, instead of defaulting to `NEXT_CACHE_KV`.
 
 ### Patch Changes
