+++ packages/cloudflare/src/api/overrides/queue/do-queue.ts
@@ -1,12 +1,12 @@
 import type { Queue, QueueMessage } from "@opennextjs/aws/types/overrides";
 import { IgnorableError } from "@opennextjs/aws/utils/error.js";
 
-import { getCloudflareContext } from "./cloudflare-context";
+import { getCloudflareContext } from "../../cloudflare-context";
 
 export default {
-  name: "durable-queue",
+  name: "do-queue",
   send: async (msg: QueueMessage) => {
-    const durableObject = getCloudflareContext().env.NEXT_CACHE_REVALIDATION_DURABLE_OBJECT;
+    const durableObject = getCloudflareContext().env.NEXT_CACHE_DO_REVALIDATION;
     if (!durableObject) throw new IgnorableError("No durable object binding for cache revalidation");
 
     const id = durableObject.idFromName(msg.MessageGroupId);
