+++ packages/cloudflare/src/api/overrides/queue/memory-queue.spec.ts
@@ -6,7 +6,7 @@ import cache, { DEFAULT_REVALIDATION_TIMEOUT_MS } from "./memory-queue.js";
 vi.mock("./.next/prerender-manifest.json", () => Promise.resolve({ preview: { previewModeId: "id" } }));
 
 const mockServiceWorkerFetch = vi.fn();
-vi.mock("./cloudflare-context", () => ({
+vi.mock("../../cloudflare-context", () => ({
   getCloudflareContext: () => ({
     env: { NEXT_CACHE_REVALIDATION_WORKER: { fetch: mockServiceWorkerFetch } },
   }),
