+++ packages/cloudflare/src/api/overrides/tag-cache/d1-tag-cache.ts
@@ -3,8 +3,8 @@ import type { OpenNextConfig } from "@opennextjs/aws/types/open-next.js";
 import type { OriginalTagCache } from "@opennextjs/aws/types/overrides.js";
 import { RecoverableError } from "@opennextjs/aws/utils/error.js";
 
-import { getCloudflareContext } from "./cloudflare-context.js";
-import { DEFAULT_NEXT_CACHE_D1_REVALIDATIONS_TABLE } from "./constants.js";
+import { getCloudflareContext } from "../../cloudflare-context.js";
+import { DEFAULT_NEXT_CACHE_D1_REVALIDATIONS_TABLE, DEFAULT_NEXT_CACHE_D1_TAGS_TABLE } from "./internal.js";
 
 /**
  * An instance of the Tag Cache that uses a D1 binding (`NEXT_CACHE_D1`) as it's underlying data store.
@@ -154,7 +154,7 @@ class D1TagCache implements OriginalTagCache {
       isDisabled: false as const,
       db,
       tables: {
-        tags: cfEnv.NEXT_CACHE_D1_TAGS_TABLE ?? "tags",
+        tags: cfEnv.NEXT_CACHE_D1_TAGS_TABLE ?? DEFAULT_NEXT_CACHE_D1_TAGS_TABLE,
         revalidations: cfEnv.NEXT_CACHE_D1_REVALIDATIONS_TABLE ?? DEFAULT_NEXT_CACHE_D1_REVALIDATIONS_TABLE,
       },
     };
