+++ packages/cloudflare/src/api/overrides/tag-cache/internal.ts
@@ -1,1 +1,2 @@
 export const DEFAULT_NEXT_CACHE_D1_REVALIDATIONS_TABLE = "revalidations";
+export const DEFAULT_NEXT_CACHE_D1_TAGS_TABLE = "tags";
