+++ packages/cloudflare/templates/open-next.config.ts
@@ -1,6 +1,6 @@
 // default open-next.config.ts file created by @opennextjs/cloudflare
 import { defineCloudflareConfig } from "@opennextjs/cloudflare/config";
-import kvIncrementalCache from "@opennextjs/cloudflare/kv-cache";
+import kvIncrementalCache from "@opennextjs/cloudflare/overrides/incremental-cache/kv-cache";
 
 export default defineCloudflareConfig({
   incrementalCache: kvIncrementalCache,
